calculator 1.02
irrelevant files removed
