+++ makefile
@@ -1,3 +1,2 @@
 all:
-	g++ -std=c++14 -Wall -Werror -O2 -o new_calculator new_calculator.cpp
 	g++ -std=c++14 -Wall -Werror -O2 -o calculator calculator.cpp
