--- out
@@ -1,12 +1,0 @@
-20 - 10 = 10
-3 + 4 = 7
-2 * 7 = 14
-sqrt 16.000 = 4.000
-sqrt 4.000 = 2.000
-2.000 * 14 = 28.000
-28.000 / 2 = 14.000
-3 * 14.000 = 42.000
-42.000 / 2 = 21.000
-3 * 21.000 = 63.000
-63.000 / 2 = 31.500
-3 * 31.500 = 94.500
